Move RangeArgument and Bound to libcore.
